# openspec/openapi-v1.yaml
openapi: 3.1.0
info:
  title: AstrBot OpenAPI v1
  version: 0.1.0
  description: >
    Target REST contract for migrating AstrBot HTTP APIs to /api/v1.
    Dynamic AstrBot configuration payloads are intentionally modeled as open
    JSON objects because their schemas are provided at runtime by template
    endpoints.
    Developer API keys currently support these scopes only: bot, provider,
    persona, im, config, chat, data, file, plugin, mcp, skill, plus the sensitive
    config:edit_admin and chat:admin sub-scopes. The config scope also grants bot
    and provider access, but no parent scope grants a sensitive sub-scope.
...
servers:
  - url: http://localhost:6185
    description: Local AstrBot server
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/analysis/customer-names:
    get:
      tags: [Analysis]
      summary: Resolve WeChat nicknames for external customer IDs
      operationId: getAnalysisCustomerNames
      x-astrbot-scope: data
      parameters:
        - name: ids
          in: query
          required: true
          schema:
            type: string
          description: Comma-separated external customer IDs.
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/analysis/scenarios:
    get:
      tags: [Analysis]
      summary: List analysis scenario prompt templates
      operationId: listAnalysisScenarios
      x-astrbot-scope: data
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    post:
      tags: [Analysis]
      summary: Create a custom analysis scenario
      operationId: createAnalysisScenario
      x-astrbot-scope: data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalysisScenarioCreateRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/analysis/scenarios/{scenario_id}/restore:
    post:
      tags: [Analysis]
      summary: Restore a built-in analysis scenario to its default content
      operationId: restoreAnalysisScenario
      x-astrbot-scope: data
      parameters:
        - name: scenario_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/analysis/scenarios/{scenario_id}:
    put:
      tags: [Analysis]
      summary: Update an analysis scenario
      operationId: updateAnalysisScenario
      x-astrbot-scope: data
      parameters:
        - name: scenario_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalysisScenarioUpsertRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Analysis]
      summary: Delete a custom analysis scenario
      operationId: deleteAnalysisScenario
      x-astrbot-scope: data
      parameters:
        - name: scenario_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/analysis/sessions:
    post:
      tags: [Analysis]
      summary: Create a pre-seeded AI analysis chat session
      operationId: createAnalysisSession
      x-astrbot-scope: data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalysisSessionRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/analysis/export:
    post:
      tags: [Analysis]
      summary: Export conversations matching a filter as CSV or JSONL
      operationId: exportAnalysisConversations
      x-astrbot-scope: data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalysisExportRequest"
      responses:
        "200":
          description: Exported conversation data
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Ok:
      description: Standard AstrBot success response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessEnvelope"
...
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          const: ok
        message:
          type: string
        data: {}
      additionalProperties: false
...
    ConversationRef:
      type: object
      required: [user_id, cid]
      properties:
        user_id:
          type: string
        cid:
          type: string
      additionalProperties: false
...
    AnalysisScenarioUpsertRequest:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        description:
          type: string
        instruction:
          type: string
      additionalProperties: false
    AnalysisScenarioCreateRequest:
      type: object
      required: [name, instruction]
      properties:
        name:
          type: string
        icon:
          type: string
        description:
          type: string
        instruction:
          type: string
      additionalProperties: false
    AnalysisSessionRequest:
      type: object
      properties:
        platforms:
          type: array
          items:
            type: string
          description: Platform IDs; defaults to wecom when omitted.
        keyword:
          type: string
        message_types:
          type: array
          items:
            type: string
          description: Message types; defaults to FriendMessage when omitted.
        created_after:
          type: integer
          description: Unix timestamp (seconds) lower bound on conversation creation.
        created_before:
          type: integer
          description: Unix timestamp (seconds) upper bound on conversation creation.
        conversations:
          type: array
          description: Explicit conversation list; overrides the filter when provided.
          items:
            $ref: "#/components/schemas/ConversationRef"
        scenario_id:
          type: string
          default: intent
      additionalProperties: false
    AnalysisExportRequest:
      allOf:
        - $ref: "#/components/schemas/AnalysisSessionRequest"
        - type: object
          properties:
            format:
              type: string
              enum: [csv, jsonl]
              default: csv
...